Make FaunaDB an option for cache and user storage
Allows user storage to attach database access token or other secrets
to the Authorization Token. The authorizer converts these to API
Gateway context, so they are available to authorized functions.

In the case of FaunaDB we use this payload to deliver a FaunaDB access
secret that corresponds to the authenticated user.